Remove Makefile dependency for shell scripts.
The Makefile dependency was considered useful when using --prefix
option with configure after a previous installation, since then, path
of executables needed to be revised. However, it has been found that
libtool "rpath" variable is not corrected and a correct package
rebuild requires the execution of "make clean". Because of this,
the makefile dependency has been removed.